Relocate common concurrency features into a separate library for reuse
