+++ Sources/AuthFoundation/JWT/JWK.swift
@@ -11,6 +11,7 @@
 //
 
 import Foundation
+import CommonSupport
 
 /// Describes an individual key from an authorization server, which can be used to validate tokens or encrypt content.
 ///
