+++ Sources/AuthFoundation/OAuth2/OAuth2Client.swift
@@ -11,6 +11,7 @@
 //
 
 import Foundation
+import CommonSupport
 
 #if canImport(FoundationNetworking)
 import FoundationNetworking
