+++ Sources/AuthFoundation/Token Management/Token.swift
@@ -11,6 +11,7 @@
 //
 
 import Foundation
+import CommonSupport
 
 /// Token information representing a user's access to a resource server, including access token, refresh token, and other related information.
 public struct Token: Sendable, Codable, Equatable, Hashable, JSONClaimContainer, Expires {
