+++ Sources/BrowserSignin/Providers/AuthenticationServicesProvider.swift
@@ -11,6 +11,7 @@
 //
 
 import AuthFoundation
+import CommonSupport
 import OAuth2Auth
 
 #if canImport(AuthenticationServices)
