+++ Sources/CommonSupport/CoalescedResult.swift
