+++ Sources/CommonSupport/LockedValue.swift
