+++ Sources/CommonSupport/WeakCollection.swift
