+++ Sources/OAuth2Auth/Authentication/ResourceOwnerFlow.swift
@@ -11,6 +11,7 @@
 //
 
 import Foundation
+import CommonSupport
 import AuthFoundation
 
 /// An authentication flow class that implements the Resource Owner Flow exchange.
