+++ Sources/OAuth2Auth/Authentication/TokenExchangeFlow.swift
@@ -10,8 +10,9 @@
 // See the License for the specific language governing permissions and limitations under the License.
 //
 
-import AuthFoundation
 import Foundation
+import CommonSupport
+import AuthFoundation
 
 /// An authentication flow class that implements the Token Exchange Flow.
 public actor TokenExchangeFlow: AuthenticationFlow {
