+++ Sources/OktaDirectAuth/DirectAuthFlow.swift
@@ -11,6 +11,7 @@
 //
 
 import Foundation
+import CommonSupport
 import AuthFoundation
 
 /// Delegate protocol used by ``DirectAuthenticationFlow``.
