+++ Sources/OktaIdxAuth/Capabilities/WebAuthnRegistrationCapability.swift
@@ -11,6 +11,7 @@
 //
 
 import Foundation
+import CommonSupport
 import AuthFoundation
 
 public enum WebAuthnCapabilityError: Error {
