+++ Sources/OktaIdxAuth/IDXFormField.swift
@@ -11,6 +11,7 @@
 //
 
 import Foundation
+import CommonSupport
 import AuthFoundation
 
 extension Remediation.Form {
