+++ Tests/AuthFoundationTests/TimeCoordinatorTests.swift
@@ -11,6 +11,7 @@
 //
 
 import XCTest
+import CommonSupport
 @testable import AuthFoundation
 import TestCommon
